refactor: create DownloadCsvButton
